fixing potentially fragile assignment of indices

+++ objc/ringbuffer.h
@@ -12,17 +12,17 @@
 #define __LIBPD_RING_BUFFER_H__
 
 typedef struct ring_buffer {
-    size_t size;
+    long size;
     char *buf_ptr;
-    volatile size_t write_idx;
-    volatile size_t read_idx;
+    volatile long write_idx;
+    volatile long read_idx;
 } ring_buffer;
 
-ring_buffer *rb_create(size_t size);
+ring_buffer *rb_create(long size);
 void rb_free(ring_buffer *buffer);
-size_t rb_available_to_write(ring_buffer *buffer);
-size_t rb_available_to_read(ring_buffer *buffer);
-int rb_write_to_buffer(ring_buffer *buffer, const char *src, size_t len);
-int rb_read_from_buffer(ring_buffer *buffer, char *dest, size_t len);
+long rb_available_to_write(ring_buffer *buffer);
+long rb_available_to_read(ring_buffer *buffer);
+int rb_write_to_buffer(ring_buffer *buffer, const char *src, long len);
+int rb_read_from_buffer(ring_buffer *buffer, char *dest, long len);
 
 #endif
